refactor : 파일 정리 및 변경 기존 기능 구현 코드를 makingLottoNumber에 옮겼습니다.
